% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{\%+\%}
\alias{\%+\%}
\title{Vectorised Addition}
\usage{
x \%+\% y
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
Returns a numeric vector.
}
\description{
I've no time for this.
}
\examples{
c(1, 2, 3) \%+\% c(10, 11, 12)
stopifnot(c(1, 2, 3) \%+\% c(10, 11, 12) == c(11, 13, 15))
stopifnot(all(is.na(c(NA_real_) \%+\% c(10, 11, 12))))
stopifnot(length(numeric(0) \%+\% c(10, 11, 12)) == 0)
}
