% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\docType{package}
\name{mojpakiet-package}
\alias{mojpakiet}
\alias{mojpakiet-package}
\title{mojpakiet: Moj Pierwszy Pakiet}
\description{
Opis.
}
\author{
\strong{Maintainer}: Roman Tyk \email{Roman.Tyk@pw.edu.pl}

}
