% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_knn.R
\name{impute_knn}
\alias{impute_knn}
\title{Fast missing value imputation using K-Nearest Neighbors}
\usage{
impute_knn(df, k = 5, n_threads = parallel::detectCores())
}
\arguments{
\item{df}{A data frame or a numeric matrix.}

\item{k}{An integer. The number of nearest neighbors to consider. Default is 5.}

\item{n_threads}{Number of threads to be used by OpenMP. Defaults to all available cores.}
}
\value{
A data frame or matrix of the same size, with NA values imputed.
}
\description{
This function utilizes a multithreaded C++ implementation
and a custom KD-Tree data structure to rapidly calculate the mean 
of the K-nearest neighbors for every missing value.
}
\examples{
dirty_data <- data.frame(
  X = c(1.0, 2.0, 1.5, 100.0, 105.0, 102.0),
  Y = c(2.0, 3.0, NA,  200.0, 205.0, NA)
)
clean_data <- impute_knn(dirty_data, k = 2, n_threads = 2)
print(clean_data)
}
